% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{calc_dispersion}
\alias{calc_dispersion}
\title{Calculate the dispersion parameter for a quasi-poisson model}
\usage{
calc_dispersion(y, X, beta, stratum_vector)
}
\arguments{
\item{y}{a vector of outcomes}

\item{X}{a matrix of predictors, typically the crossbasis output}

\item{beta}{a vector of coefficients}

\item{stratum_vector}{a vector describing the stratum}
}
\value{
a numeric scalar, the estimated dispersion parameter
}
\description{
Converted from STATA from armstrong 2014 supplement with LLM guidance.
Minimal error checking of these inputs at present.
}
\examples{
dispersion <- calc_dispersion(y = boston_deaths_tbl$daily_deaths,
              X = cb,
              stratum_vector = boston_deaths_tbl$strata,
              beta = coef(m_sub))
dispersion
}
