% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{calc_vcov}
\alias{calc_vcov}
\title{Calculate the variance-covariance matrix of a Conditional poisson model}
\usage{
calc_vcov(y, X, beta, stratum_vector)
}
\arguments{
\item{y}{a vector of outcomes}

\item{X}{a matrix of predictors, typically the crossbasis output}

\item{beta}{a vector of coefficients}

\item{stratum_vector}{a vector describing the stratum}
}
\value{
a numeric matrix, the variance-covariance matrix
}
\description{
https://link.springer.com/article/10.1007/s11222-005-4069-4
https://statomics.github.io/SGA2019/assets/poissonIRWLS-implemented.html#variance-covariance-matrix-of-the-model-parameters
}
\details{
and extended to multi-nomial case, with help from an LLM, confirmed it works.

here are other refs
https://doi.org/10.1111/j.2517-6161.1996.tb02079.x
https://doi.org/10.1093/biomet/68.2.563
}
\examples{
\dontrun{
# after running condPois_1stage on a single geo_unit:
result <- condPois_1stage(exposure_matrix, outcomes_tbl)
vcov_mat <- calc_vcov(
  y = result$`_`$out[[1]]$outcomes,
  X = as.matrix(result$`_`$out[[1]]$orig_basis),
  beta = result$`_`$out[[1]]$orig_coef,
  stratum_vector = result$`_`$out[[1]]$strata_vec
)
}
}
