% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{condPois_1stage}
\alias{condPois_1stage}
\title{Run a conditional poisson model for a single geographic unit}
\usage{
condPois_1stage(
  exposure_matrix,
  outcomes_tbl,
  argvar = NULL,
  arglag = NULL,
  maxlag = NULL,
  min_n = 50,
  strata_min = 0,
  global_cen = NULL,
  multi_zone = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{exposure_matrix}{a matrix of exposures, with columns for lag, usually created by \code{make_exposure_matrix}}

\item{outcomes_tbl}{a data.table of outcomes, created by \code{make_outcome_table}}

\item{argvar}{a list containing the \code{argvar} components for the \code{crossbasis}}

\item{arglag}{a list containing the \code{arglag} components for the \code{crossbasis}}

\item{maxlag}{an integer of the maximum lag}

\item{min_n}{an integer describing the minimum number of cases for a single region}

\item{strata_min}{an integer describing the minimum number of cases for a single strata}

\item{global_cen}{global centering point}

\item{multi_zone}{are multiple strata being used.}

\item{verbose}{used to print crossbasis args the first time}
}
\value{
a condPois_1stage model object
}
\description{
Run a conditional poisson model for a single geographic unit
}
\examples{
library(dlnm)
library(gnm)
library(ggplot2)
library(data.table)
# create exposure matrix
exposure_columns <- list(
  "date" = "date",
 "exposure" = "tmax_C",
 "geo_unit" = "TOWN20",
 "geo_unit_grp" = "COUNTY20"
)
boston_exposure_mat <- make_exposure_matrix(
subset(ma_exposure, TOWN20 == 'BOSTON'),
exposure_columns)

# create outcome table
outcome_columns <- list(
 "date" = "date",
 "outcome" = "daily_deaths",
 "factor" = 'age_grp',
 "factor" = 'sex',
 "geo_unit" = "TOWN20",
 "geo_unit_grp" = "COUNTY20"
)
boston_deaths_tbl <- make_outcome_table(boston_deaths,  outcome_columns)

# run the model
m1 <- condPois_1stage(exposure_matrix = boston_exposure_mat,
                    outcomes_tbl = boston_deaths_tbl)
condPois_1stage
//////////////////////////////////////////////////////////////////////////
==========================================================================
IF the outcomes_tbl has a FACTOR, enter a recursive loop
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATIONS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
CREATE CROSSBASIS for this single zone

TODO: HOW TO HANDLE MULTIPLE exposures ...
      maybe this becomes a function and you have?
      well no because then arglag etc would have to be lists
      so do you want to move this into exposure ?
      no i think thats what you do, you move this into exposure
      and then you keep one main exposure and the other gets labeled
      as "control" and some of the control are cb and
      some are single vectors
      maybe exposure_col also needs to be a named list
      so you know if its a one off or a cb
      or you could do control_cb or

==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
RUN GNM
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
CROSSPRED and CROSSREDUCE
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
Make a single centered basis
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
OUTPUT OBJECTS
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
