% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{forest_plot.condPois_1stage}
\alias{forest_plot.condPois_1stage}
\title{forest_plot method for condPois_1stage}
\usage{
\method{forest_plot}{condPois_1stage}(x, ...)
}
\arguments{
\item{x}{an object of class condPois_1stage}

\item{...}{other elements passed to spatial_plot}
}
\value{
called for its side-effect (warning); returns NULL invisibly
}
\description{
forest_plot method for condPois_1stage
}
\examples{
x <- structure(list(), class = "condPois_1stage")
forest_plot(x)
}
