% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_2stage.R
\name{forest_plot.condPois_2stage_list}
\alias{forest_plot.condPois_2stage_list}
\title{forest_plot method for condPois_2stage_list}
\usage{
\method{forest_plot}{condPois_2stage_list}(x, exposure_val)
}
\arguments{
\item{x}{an object of class condPois_2stage_list}

\item{exposure_val}{exposure value at which to plot}
}
\value{
a ggplot object
}
\description{
forest_plot method for condPois_2stage_list
}
