% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{forest_plot}
\alias{forest_plot}
\title{forest_plot base class}
\usage{
forest_plot(x, ...)
}
\arguments{
\item{x}{an object to dispatch to the appropriate forest_plot S3 method}

\item{...}{further arguments passed to the method}
}
\value{
output depends on the class of x; see method-specific documentation
}
\description{
forest_plot base class
}
\examples{
\dontrun{
# after running a condPois model:
result <- condPois_2stage(exposure_matrix, outcomes_tbl)
forest_plot(result, exposure_val = 30.0)
}
}
