% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{getRR.condPois_1stage_list}
\alias{getRR.condPois_1stage_list}
\title{getRR method for condPois_1stage_list}
\usage{
\method{getRR}{condPois_1stage_list}(x)
}
\arguments{
\item{x}{an object of class condPois_1stage_list}
}
\value{
a data.table of relative risk estimates across factor levels
}
\description{
getRR method for condPois_1stage_list
}
\examples{
middlesex_deaths_tbl <- make_outcome_table(
middlesex_deaths,  outcome_columns, collapse_to = 'age_grp')

# run the model
m3 <- condPois_1stage(exposure_matrix = middlesex_exposure_mat,
                     outcomes_tbl = middlesex_deaths_tbl,
                     global_cen = 15,
                     multi_zone = TRUE,
                     verbose = 1)
getRR(m3)
}
