% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_2stage.R
\name{getRR.condPois_2stage}
\alias{getRR.condPois_2stage}
\title{getRR method for condPois_2stage}
\usage{
\method{getRR}{condPois_2stage}(x)
}
\arguments{
\item{x}{an object of class condPois_2stage}
}
\value{
a data.table of relative risk estimates
}
\description{
getRR method for condPois_2stage
}
