% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_2stage.R
\name{getRR.condPois_2stage_list}
\alias{getRR.condPois_2stage_list}
\title{getRR method for condPois_2stage_list}
\usage{
\method{getRR}{condPois_2stage_list}(x)
}
\arguments{
\item{x}{an object of class condPois_2stage_list}
}
\value{
a data.table of relative risk estimates across factor levels
}
\description{
getRR method for condPois_2stage_list
}
\examples{
\dontrun{
  # set up exposure matrix and outcome table first (see condPois_2stage example)
  model_list <- condPois_2stage(ma_exposure_matrix, ma_outcomes_tbl, global_cen = 20)
  getRR(model_list)
}
}
