% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_SB.R
\name{getRR.condPois_sb}
\alias{getRR.condPois_sb}
\title{getRR method for condPois_sb}
\usage{
\method{getRR}{condPois_sb}(x)
}
\arguments{
\item{x}{an object of class condPois_sb}
}
\value{
a data.table of relative risk estimates
}
\description{
getRR method for condPois_sb
}
\examples{
\dontrun{
  # set up exposure matrix and outcome table first (see condPois_2stage example)
  model <- condPois_sb(ma_exposure_matrix, ma_outcomes_tbl, global_cen = 20)
  getRR(model)
}
}
