% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getRR}
\alias{getRR}
\title{getRR base class}
\usage{
getRR(x, ...)
}
\arguments{
\item{x}{an object to dispatch to the appropriate getRR S3 method}

\item{...}{further arguments passed to the method}
}
\value{
a data.table of relative risk estimates; exact format depends on the class of x
}
\description{
getRR base class
}
\examples{
\dontrun{
# after running a condPois model:
result <- condPois_2stage(exposure_matrix, outcomes_tbl)
getRR(result)
}
}
