% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bayes_fcns.R
\name{getSW}
\alias{getSW}
\title{Creates the list of spatial neighbors used in the \code{condPois_sb} Leroux model}
\usage{
getSW(shp, ni, include_self = T)
}
\description{
Helper function to get neighbors
}
\examples{
\dontrun{
# after loading an sf shapefile:
SW <- getSW(shp = my_shapefile, ni = 1, include_self = FALSE)
}
}
