% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{join_exposure_covariate}
\alias{join_exposure_covariate}
\title{join exposure covariate}
\usage{
join_exposure_covariate(exposure_mat, covariate_mat)
}
\arguments{
\item{covariate_mat}{}
}
\description{
join exposure covariate
}
