% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossbasis.R
\name{make_crossbasis}
\alias{make_crossbasis}
\title{Make crossbasis object}
\usage{
make_crossbasis(
  exposure_matrix,
  outcomes_tbl,
  unique_geos,
  maxlag,
  argvar,
  arglag,
  strata_min,
  min_n,
  verbose
)
}
\arguments{
\item{verbose}{}
}
\description{
Make crossbasis object
}
