% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{make_exposure_matrix}
\alias{make_exposure_matrix}
\title{Function to clean and prepare the exposure data matrix}
\usage{
make_exposure_matrix(
  data,
  column_mapping,
  months_subset = 5:9,
  dt_by = "day",
  maxgap = 5,
  maxlag = 5,
  grp_level = FALSE,
  keep_unit_exposures = FALSE,
  exposure_is_factor = FALSE
)
}
\arguments{
\item{data}{a dataset of exposures}

\item{column_mapping}{a named list that indicates relevant columns in \code{data}. for the exposure
data table, these need to be one of: c('date', "exposure", 'geo_unit', 'geo_unit_grp')}

\item{months_subset}{the warm season months for this region, default is Northern Hemisphere's
May through September (5 through 9)}

\item{dt_by}{is it daily data, or weekly or ...}

\item{maxlag}{the number of lags for the exposure variable (default is 5)}

\item{grp_level}{whether to summarize to the group level or not (default)}

\item{keep_unit_exposures}{if grp_level is true, whether to keep original unit-level exposures}

\item{exposure_is_factor}{exposure is a factor}

\item{maxgaps}{the maximum allowable missing exposure data gap, to be passed to zoo::na.approx (default is 5)}
}
\value{
a data.table of class("exposure")
}
\description{
Function to clean and prepare the exposure data matrix
}
\examples{
exposure_columns <- list(
  "date" = "date",
  "exposure" = "tmax_C",
  "geo_unit" = "TOWN20",
  "geo_unit_grp" = "COUNTY20"
)

ma_exposure_matrix <- make_exposure_matrix(
                      subset(ma_exposure,COUNTY20 \%in\% c('MIDDLESEX', 'WORCESTER') &
                      data.table::year(date) \%in\% 2012:2015), exposure_columns)
make_exposure_matrix
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATIONS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
CREATE XGRID
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
FILL NA VALUES
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
GROUP LEVEL SUMMARY?
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
ADD LAGS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
OUTPUT
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
