% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgrid.R
\name{make_xgrid}
\alias{make_xgrid}
\title{An internal function to make the xgrid}
\usage{
make_xgrid(
  data,
  column_mapping,
  time_subset,
  dt_by = "day",
  collapse_is_spatial = FALSE,
  collapse_is_temporal = FALSE
)
}
\arguments{
\item{data}{a matrix of exposures or outcomes}

\item{column_mapping}{named list that indicates relevant columns in \code{data}.}

\item{time_subset}{the time period of interest for analysis, specified as years, months, or days
must be specified by user - no default}

\item{dt_by}{either by day or by week}

\item{collapse_is_spatial}{logical, is the collapse spatial}

\item{collapse_is_temporal}{logical, is the collapse temporal}
}
\value{
a datatable of all date and geo unit combinations
}
\description{
An internal function to make the xgrid
}
\examples{
\dontrun{
exposure_columns <- list("date" = "date",
"exposure" = "tmax_C", "geo_unit" = "TOWN20",
"geo_unit_grp" = "COUNTY20")

make_xgrid(subset(ma_exposure, TOWN20 == 'BOSTON'),
exposure_columns, time_subset = list(month = 5:9))
}
//////////////////////////////////////////////////////////////////////////
==========================================================================
GET ALL DATES
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
CREATE XGRID SKELETON

For dates, you need all the dates

For outcomes, you also need all the dates within strata
that have SOME data so you can add 0s to low data places
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
JOIN WITH GROUP DATA
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
FINALLY - JOIN WITH DATA

This shouldnt add any rows
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
