% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bayes_fcns.R
\name{nx}
\alias{nx}
\title{Title}
\usage{
nx(xx, ni, include_self = T)
}
\arguments{
\item{include_self}{}
}
\description{
Facilitates recursive calls get neighbors
}
\examples{
xx <- list(c(2, 3), c(1, 3), c(1, 2))
nx(xx, ni = 1, include_self = TRUE)
}
