% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AN.R
\name{plot.calcAN}
\alias{plot.calcAN}
\title{Plot method for calcAN}
\usage{
\method{plot}{calcAN}(x, table_type, above_MMT, spatial_sub = NULL, override_limit = FALSE)
}
\arguments{
\item{x}{an object of class calcAN}

\item{table_type}{showing the rate table "rate" or number table "num"}

\item{above_MMT}{plot attributable numbers above or below the MMT}

\item{spatial_sub}{an option argument to subset the geo_units investigated}

\item{override_limit}{override the built-in plot limit}
}
\value{
a plot of either the attributable rate or number
}
\description{
Plot method for calcAN
}
\examples{
\dontrun{
# after running calcAN on a condPois model:
an_result <- calcAN(my_model)
plot(an_result, table_type = "num", above_MMT = TRUE)
}
}
