% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AN.R
\name{plot.calcAN_list}
\alias{plot.calcAN_list}
\title{Plot method for calcAN_list}
\usage{
\method{plot}{calcAN_list}(
  x,
  table_type,
  above_MMT,
  spatial_sub = NULL,
  ordered_levels = NULL,
  override_limit = FALSE
)
}
\arguments{
\item{x}{an object of class calcAN_list}

\item{table_type}{showing the rate table "rate" or number table "num"}

\item{above_MMT}{plot attributable numbers above or below the MMT}

\item{spatial_sub}{an option argument to subset the geo_units investigated}

\item{ordered_levels}{factor levels}

\item{override_limit}{override the built-in plot limit}
}
\value{
a plot with the attributable number or rate with bars for each factor level
}
\description{
Plot method for calcAN_list
}
\examples{
\dontrun{
# after running calcAN on a condPois model with a factor:
an_result <- calcAN(my_model_list)
plot(an_result, table_type = "num", above_MMT = TRUE)
}
}
