% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{plot.condPois_1stage}
\alias{plot.condPois_1stage}
\title{Plot method for condPois_1stage}
\usage{
\method{plot}{condPois_1stage}(x, xlab = NULL, ylab = NULL, title = NULL)
}
\arguments{
\item{x}{an object of class condPois_1stage}

\item{xlab}{xlab override}

\item{ylab}{ylab override}

\item{title}{title override}
}
\value{
a ggplot object
}
\description{
Plot method for condPois_1stage
}
\examples{
# create exposure matrix
exposure_columns <- list(
 "date" = "date",
 "exposure" = "tmax_C",
 "geo_unit" = "TOWN20",
)
middlesex_exposure <- subset(ma_exposure, COUNTY20 == 'MIDDLESEX')
middlesex_exposure_mat <- make_exposure_matrix(middlesex_exposure, exposure_columns)
# create outcome table
outcome_columns <- list(
 "date" = "date",
 "outcome" = "daily_deaths",
 "factor" = 'age_grp',
 "factor" = 'sex',
 "geo_unit" = "TOWN20",
 "geo_unit_grp" = "COUNTY20"
)
middlesex_deaths   <- subset(ma_deaths, COUNTY20 == 'MIDDLESEX')
middlesex_deaths_tbl <- make_outcome_table(middlesex_deaths,  outcome_columns)
# run the model
m2 <- condPois_1stage(exposure_matrix = middlesex_exposure_mat,
outcomes_tbl = middlesex_deaths_tbl, multi_zone = TRUE,
global_cen = 15)
plot(m2)
}
