% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{plot.condPois_1stage_list}
\alias{plot.condPois_1stage_list}
\title{Plot method for condPois_1stage_list}
\usage{
\method{plot}{condPois_1stage_list}(x, xlab = NULL, ylab = NULL, title = NULL)
}
\arguments{
\item{x}{an object of class condPois_1stage_list}

\item{xlab}{xlab override}

\item{ylab}{ylab override}

\item{title}{title override}
}
\value{
a ggplot object
}
\description{
Plot method for condPois_1stage_list
}
\examples{
middlesex_deaths_tbl <- make_outcome_table(
middlesex_deaths,  outcome_columns, collapse_to = 'age_grp')

# run the model
m3 <- condPois_1stage(exposure_matrix = middlesex_exposure_mat,
                     outcomes_tbl = middlesex_deaths_tbl,
                     global_cen = 15,
                     multi_zone = TRUE,
                     verbose = 1)
plot(m3)
}
