% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{plot_lag}
\alias{plot_lag}
\title{Plot lag method for condPois_1stage}
\usage{
plot_lag(x, exposure_val)
}
\arguments{
\item{x}{an object of class condPois_1stage}

\item{xlab}{xlab override}

\item{ylab}{ylab override}

\item{title}{title override}
}
\value{
a ggplot object
}
\description{
Plot lag method for condPois_1stage
}
