% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AN.R
\name{print.calcAN_list}
\alias{print.calcAN_list}
\title{Print method for calcAN_list}
\usage{
print.calcAN_list(x)
}
\arguments{
\item{x}{an object of class calcAN_list}
}
\value{
The original object, invisibly.
}
\description{
Print method for calcAN_list
}
\examples{
x <- structure(list(a = 1, b = 2), class = "calcAN_list")
print(x)
}
