% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{print.condPois_1stage}
\alias{print.condPois_1stage}
\title{Print method for condPois_1stage}
\usage{
print.condPois_1stage(x)
}
\arguments{
\item{x}{an object of class condPois_1stage}
}
\value{
invisibly returns x
}
\description{
Print method for condPois_1stage
}
\examples{
x <- structure(list(), class = "condPois_1stage")
print(x)
}
