% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{print.condPois_1stage_list}
\alias{print.condPois_1stage_list}
\title{Print method for condPois_1stage_list}
\usage{
print.condPois_1stage_list(x)
}
\arguments{
\item{x}{an object of class condPois_1stage_list}
}
\value{
invisibly returns x
}
\description{
Print method for condPois_1stage_list
}
\examples{
x <- structure(list(a = 1, b = 2), class = "condPois_1stage_list")
print(x)
}
