% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_SB.R
\name{print.condPois_sb}
\alias{print.condPois_sb}
\title{Print method for condPois_sb}
\usage{
print.condPois_sb(x)
}
\arguments{
\item{x}{an object of class condPois_sb}
}
\value{
invisibly returns x
}
\description{
Print method for condPois_sb
}
