% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bayes_fcns.R
\name{scale_nb_components}
\alias{scale_nb_components}
\title{Compute scaling factors for neighbor list components}
\usage{
scale_nb_components(x)
}
\arguments{
\item{x}{an nb object (neighbor list)}
}
\value{
a numeric vector of scaling factors, one per connected component
}
\description{
https://github.com/stan-dev/example-models/blob/master/knitr/car-iar-poisson/nb_data_funs.R
}
\examples{
library(spdep)
coords <- matrix(c(0,0, 1,0, 0,1, 1,1), ncol = 2, byrow = TRUE)
nb <- knn2nb(knearneigh(coords, k = 2))
scale_nb_components(nb)
}
