% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_strata_value.R
\name{set_strata_value}
\alias{set_strata_value}
\title{Set the strata column of xgrid}
\usage{
set_strata_value(xgrid, column_mapping, dt_by, grp_level, keep_unit)
}
\arguments{
\item{xgrid}{the xgrid data.table}

\item{column_mapping}{column mapping vector}

\item{dt_by}{what is the date to aggregate by}

\item{grp_level}{logical, group level aggregatation}

\item{keep_unit}{logical, with or without unit level outcomes or exposures}
}
\value{
a strata vector
}
\description{
Set the strata column of xgrid
}
