% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AN.R
\name{spatial_plot.calcAN}
\alias{spatial_plot.calcAN}
\title{spatial_plot method for calcAN}
\usage{
\method{spatial_plot}{calcAN}(x, shp, table_type, above_MMT, pal = "Purples")
}
\arguments{
\item{x}{an object of class calcAN}

\item{shp}{an sf shapefile with an appropriate column at which to join}

\item{table_type}{showing the rate table "rate" or number table "num"}

\item{above_MMT}{plot attributable numbers above or below the MMT}

\item{pal}{color palette}
}
\value{
a spatial plot showing attributable numbers or rates
}
\description{
spatial_plot method for calcAN
}
\examples{
\dontrun{
# after running calcAN on a condPois model with a factor:
an_result <- calcAN(my_model)
spatial_plot(an_result, shp = my_shapefile, table_type = "num", above_MMT = TRUE)
}
}
