% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{spatial_plot.condPois_1stage}
\alias{spatial_plot.condPois_1stage}
\title{spatial_plot method for condPois_1stage}
\usage{
\method{spatial_plot}{condPois_1stage}(x, ...)
}
\arguments{
\item{x}{an object of class condPois_1stage}

\item{...}{other elements passed to spatial_plot}
}
\value{
called for its side-effect (warning); returns NULL invisibly
}
\description{
spatial_plot method for condPois_1stage
}
\examples{
x <- structure(list(), class = "condPois_1stage")
spatial_plot(x)
}
