% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{spatial_plot.condPois_1stage_list}
\alias{spatial_plot.condPois_1stage_list}
\title{spatial_plot method for condPois_1stage_list}
\usage{
\method{spatial_plot}{condPois_1stage_list}(x, ...)
}
\arguments{
\item{x}{an object of class condPois_1stage_list}

\item{...}{other elements passed to spatial_plot}
}
\value{
called for its side-effect (warning); returns NULL invisibly
}
\description{
spatial_plot method for condPois_1stage_list
}
\examples{
x <- structure(list(a = 1, b = 2), class = "condPois_1stage_list")
spatial_plot(x)
}
