% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_SB.R
\name{spatial_plot.condPois_sb_list}
\alias{spatial_plot.condPois_sb_list}
\title{spatial_plot method for condPois_sb_list}
\usage{
\method{spatial_plot}{condPois_sb_list}(x, shp, exposure_val)
}
\arguments{
\item{x}{an object of class condPois_sb_list}

\item{shp}{an sf shapefile with an appropriate column at which to join}

\item{exposure_val}{exposure value at which to plot}
}
\value{
a patchwork of ggplot objects, one per factor level
}
\description{
spatial_plot method for condPois_sb_list
}
