% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{spatial_plot}
\alias{spatial_plot}
\title{spatial_plot base class}
\usage{
spatial_plot(x, ...)
}
\arguments{
\item{...}{}
}
\description{
spatial_plot base class
}
\examples{
\dontrun{
# after running a condPois model:
result <- condPois_2stage(exposure_matrix, outcomes_tbl)
spatial_plot(result, shp = my_shapefile, exposure_val = 30.0)
}
}
