% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_us_holiday.R
\name{add_US_holiday}
\alias{add_US_holiday}
\title{Function to add US holidays}
\usage{
add_US_holiday(exposure_matrix)
}
\arguments{
\item{exposure_matrix}{}
}
\value{
an outcome table object with an additional covariate
}
\description{
Function to add US holidays
}
\examples{
exposure_columns <- list(
  "date" = "date",
  "exposure" = "tmax_C",
  "geo_unit" = "TOWN20",
  "geo_unit_grp" = "COUNTY20"
)

exp_data = subset(ma_exposure, COUNTY20 \%in\% c('MIDDLESEX', 'WORCESTER'))

exposure_matrix <- make_exposure_matrix(
  data = exp_data,
  column_mapping = exposure_columns,
  time_subset = list(year = 2012:2015)
)

exposure_matrix <- add_us_holiday(exposure_matrix)
}
