% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AN.R
\name{calc_AN}
\alias{calc_AN}
\title{Calculate attributable number and attributable rates}
\usage{
calc_AN(
  model,
  outcomes_tbl,
  pop_data,
  spatial_agg_type,
  spatial_join_col,
  by_year = FALSE,
  nsim = 300,
  verbose = 0
)
}
\arguments{
\item{model}{a model object of class \code{condPois_2stage}, \code{condPois_1stage}, or \code{condPois_sb}}

\item{outcomes_tbl}{a table of outcomes, of class \code{outcomes}}

\item{pop_data}{population data}

\item{spatial_agg_type}{what is the spatial resolution you are aggregating to}

\item{spatial_join_col}{how should you join population data to the outcome table}

\item{by_year}{Export annual counts}

\item{nsim}{number of simulations required for calculation of empirical CI (default = 300)}

\item{verbose}{0 = no printing, 1 = headers, 2 = detailed}
}
\description{
Calculate attributable number and attributable rates
}
\examples{
//////////////////////////////////////////////////////////////////////////
==========================================================================
RECURSIVE CALL IF FACTOR
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATIONS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
Part 1
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
Part 2
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
OUTPUT
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
