% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argvar.R
\name{check_argvar}
\alias{check_argvar}
\title{Internal function to check argvar}
\usage{
check_argvar(argvar, this_exp)
}
\arguments{
\item{argvar}{the argvar argument that is passed in}

\item{this_exp}{the exposure vector}
}
\value{
A validated version of \code{argvar}, potentially coerced into
a standard form (e.g., a symbol or expression). Throws an error
if validation fails.
}
\description{
Internal function to check argvar
}
\details{
this function is not meant to be run directly by the user
}
\examples{
set.seed(123)
argvar_raw = NULL      #
this_exp = rnorm(100)  # some random vector of exposures

check_argvar(argvar_raw, this_exp)
}
