% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityClimateHealth-package.R
\docType{package}
\name{cityClimateHealth-package}
\alias{cityClimateHealth-package}
\alias{cityClimateHealth}
\title{The 'cityClimateHealth' package.}
\description{
Tools for estimating heat-related health impacts using daily exposure and
outcome data at small spatial scales. The package supports one-stage
conditional Poisson models, two-stage meta-analytic designs, and spatial
Bayesian approaches that borrow strength across neighboring geographic
units when case counts are small. Methods include distributed lag
non-linear models and attributable number calculations, with workflows
designed for messy real-world epidemiologic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://climatehealth.city}
}

}
\author{
\strong{Maintainer}: Chad Milando \email{cmilando@bu.edu}

Authors:
\itemize{
  \item Chad Milando \email{cmilando@bu.edu}
  \item Emma Gause \email{egause@bu.edu}
  \item Quinn Adams \email{qadams@bu.edu}
  \item Zachary Popp \email{zpopp@bu.edu}
  \item Alexis Arlak \email{aarlak@bu.edu}
  \item Gregory Wellenius \email{wellenius@bu.edu}
}

}
