% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_data}
\alias{collapse_data}
\title{collapse the data based on factors and group level strata}
\usage{
collapse_data(data, column_mapping, fcn, data_type, grp_level, keep_unit)
}
\arguments{
\item{keep_unit}{}
}
\description{
collapse the data based on factors and group level strata
}
