% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_1stage.R
\name{condPois_1stage}
\alias{condPois_1stage}
\title{Run a conditional poisson model for a single geographic unit}
\usage{
condPois_1stage(
  exposure_matrix,
  outcomes_tbl,
  argvar = NULL,
  arglag = NULL,
  maxlag = NULL,
  min_n = 50,
  strata_min = 0,
  global_cen = NULL,
  multi_zone = FALSE,
  verbose = TRUE,
  truncate = 0.025
)
}
\arguments{
\item{exposure_matrix}{a matrix of exposures, with columns for lag, usually created by \code{make_exposure_matrix}}

\item{outcomes_tbl}{a data.table of outcomes, created by \code{make_outcome_table}}

\item{argvar}{a list containing the \code{argvar} components for the \code{crossbasis}}

\item{arglag}{a list containing the \code{arglag} components for the \code{crossbasis}}

\item{maxlag}{an integer of the maximum lag}

\item{min_n}{an integer describing the minimum number of cases for a single geo_unit}

\item{strata_min}{an integer describing the minimum number of cases for a single strata}

\item{global_cen}{global centering point}

\item{multi_zone}{are multiple strata being used.}

\item{verbose}{used to print crossbasis args the first time}
}
\value{
a condPois_1stage model object
}
\description{
Run a conditional poisson model for a single geographic unit
}
\examples{
condPois_1stage
//////////////////////////////////////////////////////////////////////////
==========================================================================
IF the outcomes_tbl has a FACTOR, enter a recursive loop
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATIONS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
CREATE CROSSBASIS for this single zone
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
