% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_2stage.R
\name{condPois_2stage}
\alias{condPois_2stage}
\title{Run a conditional poisson model for across many geographic units}
\usage{
condPois_2stage(
  exposure_matrix,
  outcomes_tbl,
  global_cen = NULL,
  argvar = NULL,
  arglag = NULL,
  maxlag = NULL,
  min_n = NULL,
  rf = NULL,
  strata_min = 0,
  verbose = 0,
  truncate = 0.01
)
}
\arguments{
\item{exposure_matrix}{a matrix of exposures, with columns for lag, usually created by \code{make_exposure_matrix}}

\item{outcomes_tbl}{a data.table of outcomes, created by \code{make_outcome_table}}

\item{global_cen}{a global centering point}

\item{argvar}{a list containing the \code{argvar} components for the \code{crossbasis}}

\item{arglag}{a list containing the \code{arglag} components for the \code{crossbasis}}

\item{maxlag}{an integer of the maximum lag}

\item{min_n}{an integer describing the minimum number of cases for a single region}

\item{rf}{optional character string to use as the random formula in mixmeta}

\item{strata_min}{minimum number of cases per strata}

\item{verbose}{whether to print geo_units as they are run}

\item{truncate}{}
}
\value{
an object of class condPois_2stage
}
\description{
MixMeta
nope its this: https://www.sciencedirect.com/science/article/pii/S0140673614621140?via\%3Dihub
avgtmax and range and random effect for county
unclear, Gasp does both here https://github.com/gasparrini/RiskExtrapolation/blob/3caa5c46a748d6a789021554f435e53f9960d4c1/24_MetaRegression.R
and it goes counrty/city
https://github.com/gasparrini/MCC-SO2/blob/1a158b31bf83f3bbca3449b553a6570280306299/Rcode/06.nonlin.R#L47
nlmeta <- mixmeta(nlcoefall, nlvcovall, random=~1|country/city, data=cities,
control=list(showiter=T, igls.inititer=10))
basically, did Gasp 2015 (temp_mean, temp_iqr, random effect for county)
and included the hierarchicacy structure for ids within countries
https://github.com/gasparrini/2015_gasparrini_EHP_Rcodedata
https://github.com/gasparrini/2019_sera_StatMed_Rcode/blob/master/03.MultilevelMA.R
}
\details{
update as of 2025.12.21, its this https://github.com/gasparrini/Extended2stage/blob/cf82fd2f0d616210e8aa6911d2fb49a732f65795/02.multilevel.R#L68
}
\examples{
library(data.table)
exposure_columns <- list(
 "date" = "date",
 "exposure" = "tmax_C",
 "geo_unit" = "TOWN20",
 "geo_unit_grp" = "COUNTY20"
)

ma_exposure_matrix <- make_exposure_matrix(
 subset(ma_exposure,COUNTY20 \%in\% c('MIDDLESEX', 'WORCESTER') &
          year(date) \%in\% 2012:2015), exposure_columns)

outcome_columns <- list(
 "date" = "date",
 "outcome" = "daily_deaths",
 "factor" = 'age_grp',
 "factor" = 'sex',
 "geo_unit" = "TOWN20",
 "geo_unit_grp" = "COUNTY20"
)
ma_outcomes_tbl <- make_outcome_table(
 subset(ma_deaths,COUNTY20 \%in\% c('MIDDLESEX', 'WORCESTER') &
          year(date) \%in\% 2012:2015), outcome_columns)
ma_model <- condPois_2stage(ma_exposure_matrix, ma_outcomes_tbl,
           verbose = 1, global_cen = 10)
condPois_2stage
//////////////////////////////////////////////////////////////////////////
==========================================================================
IF the outcomes_tbl has a FACTOR, enter a recursive loop
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATION
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
STAGE 1
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
MIXMETA
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
detour to calculate blups and RRs for other levels
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
STAGE 2 - BLUP
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
OUTPUT
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
