% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cb_list.R
\name{expand_arglist}
\alias{expand_arglist}
\title{IN-PROGRESS functions that creates a list of cross-basis objects to try}
\usage{
expand_arglist(arglist)
}
\arguments{
\item{arglist}{}
}
\description{
IN-PROGRESS functions that creates a list of cross-basis objects to try
}
\examples{
arglist <- list(
  list(fun = "ns", df = c(3, 4)),
  list(fun = "bs", degree = c(2, 3))
)
expand_arglist(arglist)
}
