% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_2stage.R
\name{forest_plot.condPois_2stage_list}
\alias{forest_plot.condPois_2stage_list}
\title{forest_plot method for condPois_2stage_list}
\usage{
\method{forest_plot}{condPois_2stage_list}(x, exposure_val)
}
\arguments{
\item{x}{an object of class condPois_2stage_list}

\item{exposure_val}{exposure value at which to plot}
}
\value{
a ggplot object
}
\description{
forest_plot method for condPois_2stage_list
}
\examples{
\dontrun{
  # set up exposure matrix and outcome table first (see condPois_2stage example)
  model_list <- condPois_2stage(ma_exposure_matrix, ma_outcomes_tbl, global_cen = 20)
  forest_plot(model_list, exposure_val = 30.0)
}
}
