% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_SB.R
\name{forest_plot.condPois_sb}
\alias{forest_plot.condPois_sb}
\title{forest_plot method for condPois_sb}
\usage{
\method{forest_plot}{condPois_sb}(x, exposure_val)
}
\arguments{
\item{x}{an object of class condPois_sb}

\item{exposure_val}{exposure value at which to plot}
}
\value{
a ggplot object
}
\description{
forest_plot method for condPois_sb
}
