% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{forest_plot}
\alias{forest_plot}
\title{forest_plot base class}
\usage{
forest_plot(x, ...)
}
\description{
forest_plot base class
}
\examples{
\dontrun{
# after running a condPois model:
result <- condPois_2stage(exposure_matrix, outcomes_tbl)
forest_plot(result, exposure_val = 30.0)
}
}
