% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centered_cp.R
\name{get_centered_cp}
\alias{get_centered_cp}
\title{Internal function to get centered cp objects}
\usage{
get_centered_cp(
  argvar,
  xcoef,
  xvcov,
  this_exp,
  x_b,
  global_cen,
  cen,
  exposure_is_factor,
  truncate
)
}
\arguments{
\item{argvar}{the list of argvar elements used to determine the exposure dimension of \code{crossbasis}}

\item{xcoef}{the centered basis coefficients that resulted}

\item{xvcov}{the variance covariance matrix for the centered basis coefficients}

\item{this_exp}{a vector for the exposure}

\item{x_b}{a list that is the boundary of this_exp}

\item{global_cen}{a global centering point, if it exists}

\item{cen}{the local centering point}

\item{exposure_is_factor}{logical, if exposure is a factor}

\item{truncate}{}
}
\value{
a list object, with the centered basis and centered crosspred output
}
\description{
Needed a function for this because we do it twice: once for regional RRs
and once for BLUP
}
\examples{
\dontrun{
# after running condPois_1stage on a single geo_unit:
result <- condPois_1stage(exposure_matrix, outcomes_tbl)
centered <- get_centered_cp(
  argvar = result$`_`$out[[1]]$argvar,
  xcoef = result$`_`$out[[1]]$coef,
  xvcov = result$`_`$out[[1]]$vcov,
  this_exp = result$`_`$out[[1]]$this_exp,
  x_b = c(0, 40),
  global_cen = NULL,
  cen = result$`_`$out[[1]]$cen,
  exposure_is_factor = FALSE
)
}
}
