% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_SB.R
\name{getRR.condPois_sb}
\alias{getRR.condPois_sb}
\title{getRR method for condPois_sb}
\usage{
\method{getRR}{condPois_sb}(x)
}
\arguments{
\item{x}{an object of class condPois_sb}
}
\value{
a data.table of relative risk estimates
}
\description{
getRR method for condPois_sb
}
