% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_SB.R
\name{getRR.condPois_sb_list}
\alias{getRR.condPois_sb_list}
\title{getRR method for condPois_sb_list}
\usage{
\method{getRR}{condPois_sb_list}(x)
}
\arguments{
\item{x}{an object of class condPois_sb_list}
}
\value{
a data.table of relative risk estimates across factor levels
}
\description{
getRR method for condPois_sb_list
}
