% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getRR}
\alias{getRR}
\title{getRR base class}
\usage{
getRR(x, ...)
}
\description{
getRR base class
}
\examples{
\dontrun{
# after running a condPois model:
result <- condPois_2stage(exposure_matrix, outcomes_tbl)
getRR(result)
}
}
