% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{input_validation}
\alias{input_validation}
\title{Input validation}
\usage{
input_validation(exposure_matrix, outcomes_tbl)
}
\description{
Input validation
}
\examples{
\dontrun{
# after creating exposure and outcome objects:
validated <- input_validation(exposure_matrix, outcomes_tbl)
exposure_matrix <- validated$exposure_matrix
outcomes_tbl <- validated$outcomes_tbl
}
}
