% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{make_outcome_table}
\alias{make_outcome_table}
\title{Function to create the outcome table
TO DO : EDIT XGRID}
\usage{
make_outcome_table(
  data,
  column_mapping,
  time_subset,
  dt_by = "day",
  collapse_to = NULL,
  collapse_is_spatial = FALSE,
  collapse_is_temporal = FALSE,
  grp_level = FALSE,
  keep_unit_outcomes = FALSE
)
}
\arguments{
\item{data}{a dataset of outcomes}

\item{column_mapping}{a named list that indicates relevant columns in \code{data}. for the exposure
data table, these need to be one of: c('date', "outcome",'factor, 'geo_unit', 'geo_unit_grp')}

\item{time_subset}{the time period of interest for analysis, specified as years, months, or days
must be specified by user - no default}

\item{dt_by}{is it daily data, or weekly or ...}

\item{collapse_to}{which factors to collapse across}

\item{collapse_is_spatial}{is collapse a spatial variable}

\item{collapse_is_temporal}{is collapse a temporal variable}

\item{grp_level}{whether to summarize to the group level or not (default)}

\item{keep_unit_outcomes}{if grp_level is true, whether to keep original unit-level outcomes}
}
\value{
a data.table of class("outcome")
}
\description{
Function to create the outcome table
TO DO : EDIT XGRID
}
\examples{
outcome_columns <- list(
  "date" = "date",
  "outcome" = "daily_deaths",
  "factor" = 'age_grp',
  "factor" = 'sex',
  "geo_unit" = "TOWN20",
  "geo_unit_grp" = "COUNTY20"
)
ma_outcomes_tbl <- make_outcome_table(
 subset(ma_deaths,COUNTY20 \%in\% c('MIDDLESEX', 'WORCESTER') &
           year(date) \%in\% 2012:2015), outcome_columns)
make_outcome_table
//////////////////////////////////////////////////////////////////////////
==========================================================================
VALIDATIONS
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
COLLAPSE AND SUMMARIZE
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
MAKE XGRID AND STRATA
==========================================================================
//////////////////////////////////////////////////////////////////////////
//////////////////////////////////////////////////////////////////////////
==========================================================================
OUTPUT
==========================================================================
//////////////////////////////////////////////////////////////////////////
}
