% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bayes_fcns.R
\name{nb2graph}
\alias{nb2graph}
\title{Title}
\usage{
nb2graph(x)
}
\arguments{
\item{x}{}
}
\description{
https://github.com/stan-dev/example-models/blob/master/knitr/car-iar-poisson/nb_data_funs.R
}
\examples{
library(spdep)
coords <- matrix(c(0,0, 1,0, 0,1, 1,1), ncol = 2, byrow = TRUE)
nb <- knn2nb(knearneigh(coords, k = 2))
nb2graph(nb)
}
