% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bayes_fcns.R
\name{nb2subgraph}
\alias{nb2subgraph}
\title{Convert a neighbor list subgraph to node pairs}
\usage{
nb2subgraph(x, c_id, comp_ids, offsets)
}
\arguments{
\item{x}{an nb object (neighbor list)}

\item{c_id}{subcomponent id to extract}

\item{comp_ids}{vector of subcomponent ids for each node}

\item{offsets}{vector of per-component node offsets}
}
\value{
a list with node1 and node2 integer vectors defining graph edges
}
\description{
https://github.com/stan-dev/example-models/blob/master/knitr/car-iar-poisson/nb_data_funs.R
}
\examples{
library(spdep)
coords <- matrix(c(0,0, 1,0, 0,1, 1,1), ncol = 2, byrow = TRUE)
nb <- knn2nb(knearneigh(coords, k = 2))
comp_ids <- n.comp.nb(nb)[[2]]
offsets <- indexByComponent(comp_ids)
nb2subgraph(nb, c_id = 1, comp_ids = comp_ids, offsets = offsets)
}
