% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_2stage.R
\name{plot.condPois_2stage}
\alias{plot.condPois_2stage}
\title{Plot method for condPois_2stage}
\usage{
\method{plot}{condPois_2stage}(x, geo_unit, xlab = NULL, ylab = NULL, title = NULL)
}
\arguments{
\item{x}{an object of class condPois_2stage}

\item{geo_unit}{a geo_unit to investigate}

\item{xlab}{xlab override}

\item{ylab}{ylab override}

\item{title}{title override}
}
\value{
a ggplot object
}
\description{
Plot method for condPois_2stage
}
\examples{
\dontrun{
  # set up exposure matrix and outcome table first (see condPois_2stage example)
  model <- condPois_2stage(ma_exposure_matrix, ma_outcomes_tbl, global_cen = 20)
  plot(model, geo_unit = "BOSTON")
}
}
