% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_SB.R
\name{plot.condPois_sb_list}
\alias{plot.condPois_sb_list}
\title{Plot method for condPois_sb_list}
\usage{
\method{plot}{condPois_sb_list}(x, geo_unit, xlab = NULL, ylab = NULL, title = NULL)
}
\arguments{
\item{x}{an object of class condPois_sb_list}

\item{geo_unit}{a geo_unit to investigate}

\item{xlab}{xlab override}

\item{ylab}{ylab override}

\item{title}{title override}
}
\value{
a ggplot object
}
\description{
Plot method for condPois_sb_list
}
\examples{
\dontrun{
  # set up exposure matrix and outcome table first (see condPois_2stage example)
  model_list <- condPois_sb(ma_exposure_matrix, ma_outcomes_tbl, global_cen = 20)
  plot(model_list, geo_unit = "BOSTON")
}
}
