% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_2stage.R
\name{print.condPois_2stage}
\alias{print.condPois_2stage}
\title{Print method for condPois_2stage}
\usage{
print.condPois_2stage(x)
}
\arguments{
\item{x}{an object of class condPois_2stage}
}
\value{
invisibly returns x
}
\description{
Print method for condPois_2stage
}
