% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_SB.R
\name{print.condPois_sb_list}
\alias{print.condPois_sb_list}
\title{Print method for condPois_sb_list}
\usage{
print.condPois_sb_list(x)
}
\arguments{
\item{x}{an object of class condPois_sb_list}
}
\value{
invisibly returns x
}
\description{
Print method for condPois_sb_list
}
