% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condPois_2stage.R
\name{spatial_plot.condPois_2stage}
\alias{spatial_plot.condPois_2stage}
\title{spatial_plot method for condPois_2stage}
\usage{
\method{spatial_plot}{condPois_2stage}(x, shp, exposure_val, RRlimits = NULL)
}
\arguments{
\item{x}{an object of class condPois_2stage}

\item{shp}{an sf shapefile with an appropriate column at which to join}

\item{exposure_val}{exposure value at which to plot}
}
\value{
a ggplot object
}
\description{
spatial_plot method for condPois_2stage
}
\examples{
\dontrun{
  # set up exposure matrix and outcome table first (see condPois_2stage example)
  model <- condPois_2stage(ma_exposure_matrix, ma_outcomes_tbl, global_cen = 20)
  spatial_plot(model, shp = ma_shp, exposure_val = 30.0)
}
}
