% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_subset_validate.R
\name{time_subset_validate}
\alias{time_subset_validate}
\title{validation of the time_subset argumetn}
\usage{
time_subset_validate(time_subset)
}
\arguments{
\item{time_subset}{a time_subset input}
}
\value{
a cleaned time_subset
}
\description{
validation of the time_subset argumetn
}
